updating the base model
the base model now has a structure that is general enough that it
can be used in many places and at the same time, most of the word
is done within the base model class. the two methods that are used
outside of it are `parse_base_data` and `create`, which are where
most of the modifications happen. `parse_base_data` converts a list
of sequences and labels (all mit confessions labels) into just numpy
arrays and `create` creates and stores the keras.Model in place
within the instance of BaseModel.